refactoring: split repodb.go
